% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmstools-package.R
\docType{package}
\name{brmstools-package}
\alias{brmstools}
\alias{brmstools-package}
\title{brmstools: Tools and Helpers for brms Package}
\description{
Postprocessing and plotting tools for brmsfit objects
  from brms package.
}
\author{
\strong{Maintainer}: Matti Vuorre \email{mv2521@columbia.edu} (0000-0001-5052-066X)

}
\keyword{internal}
