% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{coefplot}
\alias{coefplot}
\title{Coefficient plots for brms models}
\usage{
coefplot(model, grouping = NA, level = 0.95, pars = NA, type = "br",
  interval_col = "black", interval_h = 0.1, point_col = "black",
  point_size = 3, r_col = "black", r_intervals = FALSE, r_alpha = 0.5,
  ...)
}
\arguments{
\item{model}{A brmsfit model}

\item{grouping}{Name of grouping variable (e.g. \code{(1 | grouping)}). Defaults
to \code{NA} which returns the unique / first grouping factor in model.}

\item{level}{For limits of credible intervals}

\item{pars}{Which parameters to show (defaults to NA (all parameters)).}

\item{type}{Show population level ("b"), group-specific ("r"), or
both types ("br", default) of parameters.}

\item{interval_col}{Color of intervals}

\item{interval_h}{Height of bars at credible interval limits}

\item{point_col}{Color of points}

\item{point_size}{Size of points}

\item{r_col}{Color of points for group-specific posterior means}

\item{r_intervals}{Show credible intervals for group-specific parameters?}

\item{r_alpha}{Transparency of group-specific estimates}

\item{...}{Passed to \code{\link[=tidyfitted]{tidyfitted()}}}
}
\value{
a ggplot
}
\description{
A more traditional version of \code{\link[bayesplot:mcmc_intervals]{bayesplot::mcmc_intervals()}}, additionally
showing parameter means (and CIs) for each grouping factor.
}
