% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{Forest plots for brmsfit objects}
\usage{
forest(model, level = 0.95, sort = TRUE, show_data = FALSE,
  col_ridge = NA, density = TRUE, text = TRUE, rel_min_height = 0.01,
  scale = 0.9, fill_ridge = "grey75", digits = 2)
}
\arguments{
\item{model}{A brmsfit object.}

\item{level}{The "Confidence" level for the Credible Intervals.
Defaults to 0.95.}

\item{sort}{Logical; whether to sort the estimates in ascending
order of magnitude from bottom to top. Defaults to FALSE.}

\item{show_data}{Logical; whether to show the observed effect size
and standard error below the meta-analytic estimates. Defaults to FALSE.}

\item{col_ridge}{String; color for the outlines of the densities. Default: NA.}

\item{density}{Logical; should densities be shown?}

\item{text}{Logical; should estimates' numerical summaries be included?}

\item{rel_min_height}{Passed to [ggridges::geom_density_ridges()].}

\item{scale}{Passed to [ggridges::geom_density_ridges()].}

\item{fill_ridge}{String; color to fill the densities. Defaults to "grey65".}

\item{digits}{Digits to display in numerical summaries.}
}
\description{
This function draws a forest plot for a brmsfit model;
especially useful for random-effects meta-analytic models.
}
