% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_grouping}
\alias{get_grouping}
\title{Get grouping factors for brmsfit model}
\usage{
get_grouping(model, grouping)
}
\arguments{
\item{model}{a brmsfit model}

\item{grouping}{name of grouping factor or \code{NA}.}
}
\value{
character
}
\description{
Used internally by brmstools
}
