% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panels.R
\name{panels}
\alias{panels}
\title{Panel plots for brmsfit models}
\usage{
panels(model, grouping = NA, level = 0.95, data = T, sort = NA,
  xvar = NA, line_col = "black", line_lty = 1, ribbon_col = NA,
  ribbon_fill = "grey60", ribbon_alpha = 0.2, ...)
}
\arguments{
\item{model}{A brmsfit model}

\item{grouping}{Name of grouping variable (e.g. \code{(1 | grouping)}). Defaults
to \code{NA} which returns the unique / first grouping factor in model.}

\item{level}{For limits of credible intervals}

\item{data}{Should data be shown?}

\item{sort}{Sort panels based on a variable}

\item{xvar}{What variable should be on the x-axis}

\item{line_col}{Line color}

\item{line_lty}{Line type}

\item{ribbon_col}{Ribbon color}

\item{ribbon_fill}{Ribbon fill}

\item{ribbon_alpha}{Ribbon transparency}

\item{...}{Passed to \code{\link[=tidyfitted]{tidyfitted()}}}
}
\value{
a ggplot
}
\description{
Panel plots display fitted values for each grouping factor in a
separate panel.
}
