% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti.R
\name{spaghetti}
\alias{spaghetti}
\title{Spaghetti plots for brmsfit objects}
\usage{
spaghetti(model, grouping = NA, level = 0.95, xvar = NA,
  line_col = "black", line_lty = 1, ribbon_col = NA,
  ribbon_fill = "grey60", ribbon_alpha = 0.3, average = T, ...)
}
\arguments{
\item{model}{A brmsfit model}

\item{grouping}{Name of grouping variable (e.g. \code{(1 | grouping)}). Defaults
to \code{NA} which returns the unique / first grouping factor in model.}

\item{level}{For limits of credible intervals}

\item{xvar}{What variable should be on the x-axis}

\item{line_col}{Line color}

\item{line_lty}{Line type}

\item{ribbon_col}{Ribbon color}

\item{ribbon_fill}{Ribbon fill}

\item{ribbon_alpha}{Ribbon transparency}

\item{average}{Show average fitted values?}

\item{...}{Passed to \code{\link[=tidyfitted]{tidyfitted()}}}
}
\value{
a ggplot
}
\description{
Spaghetti plots display fitted values for each grouping factor
in a single panel with (optional) the average fitted line.
}
