% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfitted.R
\name{tidyfitted}
\alias{tidyfitted}
\title{Tidy fitted values from brmsfit models}
\usage{
tidyfitted(model, xvar = NA, level = 0.95, ...)
}
\arguments{
\item{model}{a brmsfit model}

\item{xvar}{Predictor variable to evaluate fitted values on}

\item{level}{For credible interval limits.}

\item{...}{Passed to \code{\link[brms:fitted.brmsfit]{brms::fitted.brmsfit()}}}
}
\value{
a tibble
}
\description{
This function extracts fitted values in a format expected by other
plotting functions
}
