% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{Forest plots for brmsfit objects}
\usage{
forest(model, grouping = NA, pars = NA, level = 0.95,
  av_name = "Average", sort = TRUE, show_data = FALSE,
  col_ridge = NA, fill_ridge = "grey75", density = TRUE,
  text = TRUE, rel_min_height = 0.01, scale = 0.9, digits = 2,
  theme_forest = TRUE)
}
\arguments{
\item{model}{A brmsfit object.}

\item{grouping}{Name of grouping variable (e.g. \code{(1 | grouping)}). Defaults
to \code{NA} which returns the unique / first grouping factor in model.}

\item{pars}{Parameters to plot, defaults to all (NA).}

\item{level}{The "Confidence" level for the Credible Intervals.
Defaults to 0.95.}

\item{av_name}{Name of average parameter (e.g. \code{"Meta-Analytic\\nestimate"})}

\item{sort}{Logical; whether to sort the estimates in ascending
order of magnitude from bottom to top. Defaults to FALSE.}

\item{show_data}{Logical; whether to show the observed effect size
and standard error below the meta-analytic estimates. Defaults to FALSE.}

\item{col_ridge}{String; color for the outlines of the densities. Default: NA.}

\item{fill_ridge}{String; color to fill the densities. Defaults to "grey65".}

\item{density}{Logical; should densities be shown?}

\item{text}{Logical; should estimates' numerical summaries be included?}

\item{rel_min_height}{Passed to \code{\link[ggridges:geom_density_ridges]{ggridges::geom_density_ridges()}}.}

\item{scale}{Passed to \code{\link[ggridges:geom_density_ridges]{ggridges::geom_density_ridges()}}.}

\item{digits}{Digits to display in numerical summaries.}

\item{theme_forest}{Use \code{\link[=theme_forest]{theme_forest()}} ggplot2 theme?}
}
\value{
a ggplot
}
\description{
This function draws a forest plot for a brmsfit model;
especially useful for random-effects meta-analytic models.
}
