% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycoef.R
\name{tidycoef}
\alias{tidycoef}
\title{Tidy parameters from brmsfit objects}
\usage{
tidycoef(model, summary = FALSE, level = 0.95)
}
\arguments{
\item{level}{}
}
\value{
a tibble
}
\description{
Extract brmsfit parameter (summaries) into a tidy tibble. Returns a tibble
where the type of parameter (varying / average) is represented separately
from the parameter. Especially useful for plots where the "fixed"
parameters should be shown next to their "random" counterparts.
}
